Split out EmbedderTest{Context,Compositor} to handle software and GL separately (#20962)
