+++ shell/platform/embedder/BUILD.gn
@@ -141,10 +141,20 @@ if (current_toolchain == host_toolchain) {
       "tests/embedder_config_builder.h",
       "tests/embedder_test.cc",
       "tests/embedder_test.h",
+      "tests/embedder_test_backingstore_producer.cc",
+      "tests/embedder_test_backingstore_producer.h",
       "tests/embedder_test_compositor.cc",
       "tests/embedder_test_compositor.h",
+      "tests/embedder_test_compositor_gl.cc",
+      "tests/embedder_test_compositor_gl.h",
+      "tests/embedder_test_compositor_software.cc",
+      "tests/embedder_test_compositor_software.h",
       "tests/embedder_test_context.cc",
       "tests/embedder_test_context.h",
+      "tests/embedder_test_context_gl.cc",
+      "tests/embedder_test_context_gl.h",
+      "tests/embedder_test_context_software.cc",
+      "tests/embedder_test_context_software.h",
       "tests/embedder_unittests.cc",
     ]
