+++ shell/platform/embedder/tests/embedder_a11y_unittests.cc
@@ -24,7 +24,7 @@ using Embedder11yTest = testing::EmbedderTest;
 
 // TODO(52372): De-flake and re-enable.
 TEST_F(Embedder11yTest, DISABLED_A11yTreeIsConsistent) {
-  auto& context = GetEmbedderContext();
+  auto& context = GetEmbedderContext(ContextType::kOpenGLContext);
 
   fml::AutoResetWaitableEvent latch;
