+++ shell/platform/embedder/tests/embedder_config_builder.cc
@@ -27,28 +27,28 @@ EmbedderConfigBuilder::EmbedderConfigBuilder(
 
   opengl_renderer_config_.struct_size = sizeof(FlutterOpenGLRendererConfig);
   opengl_renderer_config_.make_current = [](void* context) -> bool {
-    return reinterpret_cast<EmbedderTestContext*>(context)->GLMakeCurrent();
+    return reinterpret_cast<EmbedderTestContextGL*>(context)->GLMakeCurrent();
   };
   opengl_renderer_config_.clear_current = [](void* context) -> bool {
-    return reinterpret_cast<EmbedderTestContext*>(context)->GLClearCurrent();
+    return reinterpret_cast<EmbedderTestContextGL*>(context)->GLClearCurrent();
   };
   opengl_renderer_config_.present_with_info =
       [](void* context, const FlutterPresentInfo* present_info) -> bool {
-    return reinterpret_cast<EmbedderTestContext*>(context)->GLPresent(
+    return reinterpret_cast<EmbedderTestContextGL*>(context)->GLPresent(
         present_info->fbo_id);
   };
   opengl_renderer_config_.fbo_with_frame_info_callback =
       [](void* context, const FlutterFrameInfo* frame_info) -> uint32_t {
-    return reinterpret_cast<EmbedderTestContext*>(context)->GLGetFramebuffer(
+    return reinterpret_cast<EmbedderTestContextGL*>(context)->GLGetFramebuffer(
         *frame_info);
   };
   opengl_renderer_config_.make_resource_current = [](void* context) -> bool {
-    return reinterpret_cast<EmbedderTestContext*>(context)
+    return reinterpret_cast<EmbedderTestContextGL*>(context)
         ->GLMakeResourceCurrent();
   };
   opengl_renderer_config_.gl_proc_resolver = [](void* context,
                                                 const char* name) -> void* {
-    return reinterpret_cast<EmbedderTestContext*>(context)->GLGetProcAddress(
+    return reinterpret_cast<EmbedderTestContextGL*>(context)->GLGetProcAddress(
         name);
   };
   opengl_renderer_config_.fbo_reset_after_present = true;
@@ -72,7 +72,7 @@ EmbedderConfigBuilder::EmbedderConfigBuilder(
           return false;
         }
         bitmap.setImmutable();
-        return reinterpret_cast<EmbedderTestContext*>(context)->SofwarePresent(
+        return reinterpret_cast<EmbedderTestContextSoftware*>(context)->Present(
             SkImage::MakeFromBitmap(bitmap));
       };
 
@@ -124,7 +124,7 @@ void EmbedderConfigBuilder::SetOpenGLFBOCallBack() {
     frame_info.struct_size = sizeof(FlutterFrameInfo);
     frame_info.size.width = 0;
     frame_info.size.height = 0;
-    return reinterpret_cast<EmbedderTestContext*>(context)->GLGetFramebuffer(
+    return reinterpret_cast<EmbedderTestContextGL*>(context)->GLGetFramebuffer(
         frame_info);
   };
 }
@@ -134,7 +134,7 @@ void EmbedderConfigBuilder::SetOpenGLPresentCallBack() {
   FML_CHECK(renderer_config_.type == FlutterRendererType::kOpenGL);
   renderer_config_.open_gl.present = [](void* context) -> bool {
     // passing a placeholder fbo_id.
-    return reinterpret_cast<EmbedderTestContext*>(context)->GLPresent(0);
+    return reinterpret_cast<EmbedderTestContextGL*>(context)->GLPresent(0);
   };
 }
 
@@ -267,6 +267,16 @@ void EmbedderConfigBuilder::SetCompositor() {
 
 FlutterCompositor& EmbedderConfigBuilder::GetCompositor() {
   return compositor_;
+}
+
+void EmbedderConfigBuilder::SetRenderTargetType(
+    EmbedderTestBackingStoreProducer::RenderTargetType type) {
+  auto& compositor = context_.GetCompositor();
+  // TODO(wrightgeorge): figure out a better way of plumbing through the
+  // GrDirectContext
+  compositor.SetBackingStoreProducer(
+      std::make_unique<EmbedderTestBackingStoreProducer>(
+          compositor.GetGrContext(), type));
 }
 
 UniqueEngine EmbedderConfigBuilder::LaunchEngine() const {
